Move API/main.go to api.go and start moving pieces from Server.go to api.go
